initial page load location handler invoked in VanillaJS
